PowerFungus Component
Added the component needed for the mushroom generator.

Counts the number of plants attached to the generator and generates energy based on their lifespan.

I was thinking of making it dependent on plant growth, but judging by the way the facilities (through which I did the plant count) work, this can lead to a strange bug, with a slow recalculation.